+++ src/components/app/App.tsx
@@ -1,4 +1,15 @@
-import MainPage from '../../pages/mainPage/MainPage';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { AppRoute, AuthorizationStatus } from '../../const';
+import PrivateRoute from '../../private-route/PrivateRoute';
+import { HelmetProvider } from 'react-helmet-async';
+import MainPage from '../../pages/main-page/main-page';
+import AddReview from '../../pages/add-review/add-review';
+import Film from '../../pages/film/film';
+import MyList from '../../pages/my-list/my-list';
+import NotFoundPage from '../../pages/not-found-page/not-found-page';
+import Player from '../../pages/player/player';
+import SignIn from '../../pages/sign-in/sign-in';
+
 
 type AppProps = {
   promoFilmTitle: string;
@@ -8,7 +19,26 @@ type AppProps = {
 
 function App(props: AppProps): JSX.Element {
   return (
-    <MainPage {...props} />
+    <HelmetProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path={AppRoute.Main} element={<MainPage {...props} />} />
+          <Route path={AppRoute.SignIn} element={<SignIn />} />
+          <Route path={AppRoute.Film} element={<Film />} />
+          <Route path={AppRoute.AddReview} element={<AddReview />} />
+          <Route
+            path={AppRoute.MyList}
+            element={
+              <PrivateRoute authorizationStatus={AuthorizationStatus.NoAuth}>
+                <MyList />
+              </PrivateRoute>
+            }
+          />
+          <Route path={AppRoute.Player} element={<Player />} />
+          <Route path='*' element={<NotFoundPage />} />
+        </Routes>
+      </BrowserRouter>
+    </HelmetProvider>
   );
 }
