+++ src/components/logo/logo.tsx
@@ -1,13 +1,19 @@
-function HeaderLogo(): JSX.Element {
+import { Link } from 'react-router-dom';
+
+type LogoProps = {
+  className: string;
+}
+
+function Logo({ className }: LogoProps): JSX.Element {
   return (
     <div className="logo">
-      <a className="logo__link">
+      <Link to="main.html" className={className}>
         <span className="logo__letter logo__letter--1">W</span>
         <span className="logo__letter logo__letter--2">T</span>
         <span className="logo__letter logo__letter--3">W</span>
-      </a>
+      </Link>
     </div>
   );
 }
 
-export default HeaderLogo;
+export default Logo;
