+++ src/pages/add-review/add-review.tsx
@@ -1,0 +1,80 @@
+import { Helmet } from 'react-helmet-async';
+import { Link } from 'react-router-dom';
+import Logo from '../../components/logo/logo';
+import Rating from '../../components/rating/rating';
+
+function generateRatingList(): number[] {
+  return Array.from({ length: 10 }, (_, i) => 10 - i);
+}
+const ratingList = generateRatingList();
+
+function AddReview(): JSX.Element {
+  return (
+    <div>
+      <Helmet><title>Add Review</title></Helmet>
+      <section className="film-card film-card--full">
+        <div className="film-card__header">
+          <div className="film-card__bg">
+            <img src="img/bg-the-grand-budapest-hotel.jpg" alt="The Grand Budapest Hotel"/>
+          </div>
+          <h1 className="visually-hidden">WTW</h1>
+          <header className="page-header">
+            <Logo className={'logo__link'}/>
+            <nav className="breadcrumbs">
+              <ul className="breadcrumbs__list">
+                <li className="breadcrumbs__item">
+                  <Link to="film-page.html" className="breadcrumbs__link">
+                    The Grand Budapest Hotel
+                  </Link>
+                </li>
+                <li className="breadcrumbs__item">
+                  <a className="breadcrumbs__link">Add review</a>
+                </li>
+              </ul>
+            </nav>
+            <ul className="user-block">
+              <li className="user-block__item">
+                <div className="user-block__avatar">
+                  <img src="img/avatar.jpg" alt="User avatar" width="63" height="63"/>
+                </div>
+              </li>
+              <li className="user-block__item">
+                <a className="user-block__link">Sign out</a>
+              </li>
+            </ul>
+          </header>
+          <div className="film-card__poster film-card__poster--small">
+            <img
+              src="img/the-grand-budapest-hotel-poster.jpg"
+              alt="The Grand Budapest Hotel poster"
+              width="218"
+              height="327"
+            />
+          </div>
+        </div>
+        <div className="add-review">
+          <form action="#" className="add-review__form">
+            <div className="rating">
+              <div className="rating__stars">
+                {ratingList.map((value) => (<Rating key={value} num={value} />))}
+              </div>
+            </div>
+            <div className="add-review__text">
+              <textarea
+                className="add-review__textarea"
+                name="review-text"
+                id="review-text"
+                placeholder="Review text"
+              />
+              <div className="add-review__submit">
+                <button className="add-review__btn" type="submit">Post</button>
+              </div>
+            </div>
+          </form>
+        </div>
+      </section>
+    </div>
+  );
+}
+
+export default AddReview;
