+++ src/pages/not-found-page/not-found-page.tsx
@@ -1,0 +1,14 @@
+import { Helmet } from 'react-helmet-async';
+import { Link } from 'react-router-dom';
+
+function NotFoundPage(): JSX.Element {
+  return (
+    <>
+      <Helmet><title>Page Not Found</title></Helmet>
+      <h1>404. Page not found</h1>
+      <Link to="/">Вернуться на главную</Link>
+    </>
+  );
+}
+
+export default NotFoundPage;
